Fixed Ticket 31: Collision code fails with multiple volumes.
Cylinder end caps were always being treated as top caps, and so the entry/exit status of bottom cap hits was inverted (i.e. an exit was always counted as an entrance and vice-versa). This should be fixed now, and leaves a useful piece of code commented out in the intersection function for viewing the intercept list.